split libs in multiple files
