+++ include/glap/v2/discard.h
@@ -1,0 +1,6 @@
+#pragma once
+namespace glap::v2
+{
+    struct Discard {};
+    static constexpr Discard discard = {};
+}
