+++ tests/main.cpp
@@ -9,7 +9,7 @@
 struct Param1 {
     static constexpr std::string_view longname = "param1";  
     static constexpr std::optional<char32_t> shortname = 'p';
-    static constexpr auto type = glap::v2::ParameterType::Argument;
+    static constexpr auto type = glap::v2::model::ParameterType::Argument;
     using result_type = void;
 };
 
@@ -29,8 +29,8 @@ struct Print
 {};
 
 template <class Names, auto... T>
-struct Print<glap::v2::Argument<Names, T...>> {
-    using value_type= glap::v2::Argument<Names, T...>;
+struct Print<glap::v2::model::Argument<Names, T...>> {
+    using value_type= glap::v2::model::Argument<Names, T...>;
     void operator()(const value_type& v) const {
         fmt::print("    --{}: ", v.Longname);
         if (v.value) {
@@ -41,16 +41,16 @@ struct Print<glap::v2::Argument<Names, T...>> {
     }
 };
 template <class ...P>
-struct Print<glap::v2::Flag<P...>> {
-    using value_type= glap::v2::Flag<P...>;
+struct Print<glap::v2::model::Flag<P...>> {
+    using value_type= glap::v2::model::Flag<P...>;
     void operator()(const value_type& v) const {
         fmt::print("    --{}: {}x\n", v.Longname, v.occurences);
     }
 };
 template <class T> 
-    requires  requires (T a){
+    requires requires (T a){
         a.values;
-        T::type == glap::v2::ParameterType::Input;
+        T::type == glap::v2::model::ParameterType::Input;
     }
 struct Print<T> {
     using value_type= T;
@@ -63,8 +63,8 @@ struct Print<T> {
     }
 };
 template <class Names, auto N, auto ...Args>
-struct Print<glap::v2::Arguments<Names, N, Args...>> {
-    using value_type= glap::v2::Arguments<Names, N  , Args...>;
+struct Print<glap::v2::model::Arguments<Names, N, Args...>> {
+    using value_type= glap::v2::model::Arguments<Names, N  , Args...>;
     void operator()(const value_type& v) const {
         fmt::print("    --{}: ", v.Longname);
         auto nb=0;
@@ -81,7 +81,7 @@ template <class T>
 static constexpr auto print = Print<T>{};
 
 template <class Names, class ...P>
-auto print_command(glap::v2::Command<Names, P...>& command) {
+auto print_command(glap::v2::model::Command<Names, P...>& command) {
     fmt::print("command: {}\n", command.Longname);
     ([&] {
         print<P>(std::get<P>(command.params));
@@ -94,7 +94,8 @@ bool is_hello_world(std::string_view v) {
 
 int main(int argc, char** argv)
 {
-    using namespace glap::v2;
+    using namespace glap::v2::model;
+    using glap::v2::discard;
     // Parser<
     //     glap::v2::Command<glap::v2::Names<"commands", glap::v2::discard>, Param1>,
     //     glap::v2::Command<glap::v2::Names<"command", 'c'>, 
@@ -103,11 +104,11 @@ int main(int argc, char** argv)
     //         glap::v2::Argument<glap::v2::Names<"integer", 'g'>, to_int>
     //         glap::v2::Inputs<>
     // >> parser;
-    Parser<glap::v2::Command<glap::v2::Names<"othercommand", 't'>, glap::v2::Flag<glap::v2::Names<"flag", 'f'>>>,
-        glap::v2::Command<glap::v2::Names<"command", glap::v2::discard>, 
-            glap::v2::Flag<glap::v2::Names<"flag", 'f'>>,
-            glap::v2::Argument<glap::v2::Names<"arg", 'a'>, discard, is_hello_world>,
-            glap::v2::Arguments<glap::v2::Names<"args", 'b'>>,
+    glap::v2::Parser<Command<glap::v2::Names<"othercommand", 't'>, Flag<glap::v2::Names<"flag", 'f'>>>,
+        Command<glap::v2::Names<"command", glap::v2::discard>, 
+            Flag<glap::v2::Names<"flag", 'f'>>,
+            Argument<glap::v2::Names<"arg", 'a'>, discard, is_hello_world>,
+            Arguments<glap::v2::Names<"args", 'b'>>,
             Inputs<>
         >
     > parser;
